+++ etc/src/System/Etc/Internal/Resolver/Cli/Command.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE FlexibleContexts  #-}
+{-# LANGUAGE NamedFieldPuns    #-}
 {-# LANGUAGE NoImplicitPrelude #-}
 {-# LANGUAGE OverloadedStrings #-}
 module System.Etc.Internal.Resolver.Cli.Command (resolveCommandCli, resolveCommandCliPure) where
@@ -22,12 +23,13 @@ import qualified System.Etc.Internal.Spec.Types as Spec
 
 entrySpecToJsonCli
   :: (MonadThrow m)
-  => Bool
+  => Spec.ConfigValueType
+  -> Bool
   -> Spec.CliEntrySpec cmd
   -> m (Vector cmd, Opt.Parser (Maybe (Value JSON.Value)))
-entrySpecToJsonCli sensitive entrySpec = case entrySpec of
+entrySpecToJsonCli cvType isSensitive entrySpec = case entrySpec of
   Spec.CmdEntry commandJsonValue specSettings ->
-    return (commandJsonValue, settingsToJsonCli sensitive specSettings)
+    return (commandJsonValue, settingsToJsonCli cvType isSensitive specSettings)
 
   Spec.PlainEntry{} -> throwM CommandKeyMissing
 
@@ -35,10 +37,11 @@ configValueSpecToCli
   :: (MonadThrow m, Eq cmd, Hashable cmd)
   => HashMap cmd (Opt.Parser ConfigValue)
   -> Text
+  -> Spec.ConfigValueType
   -> Bool
   -> Spec.ConfigSources cmd
   -> m (HashMap cmd (Opt.Parser ConfigValue))
-configValueSpecToCli acc0 specEntryKey sensitive sources =
+configValueSpecToCli acc0 specEntryKey cvType isSensitive sources =
   let updateAccConfigOptParser configValueParser accOptParser =
         (\configValue accSubConfig -> case accSubConfig of
             ConfigValue{} -> accSubConfig
@@ -52,7 +55,7 @@ configValueSpecToCli acc0 specEntryKey sensitive sources =
         Nothing        -> return acc0
 
         Just entrySpec -> do
-          (commands, jsonOptParser) <- entrySpecToJsonCli sensitive entrySpec
+          (commands, jsonOptParser) <- entrySpecToJsonCli cvType isSensitive entrySpec
 
           let configValueParser = jsonToConfigValue <$> jsonOptParser
 
@@ -111,8 +114,8 @@ specToConfigValueCli
   -> (Text, Spec.ConfigValue cmd)
   -> m (HashMap cmd (Opt.Parser ConfigValue))
 specToConfigValueCli acc (specEntryKey, specConfigValue) = case specConfigValue of
-  Spec.ConfigValue _ sensitive sources ->
-    configValueSpecToCli acc specEntryKey sensitive sources
+  Spec.ConfigValue { Spec.configValueType, Spec.isSensitive, Spec.configSources } ->
+    configValueSpecToCli acc specEntryKey configValueType isSensitive configSources
 
   Spec.SubConfig subConfigSpec -> subConfigSpecToCli specEntryKey subConfigSpec acc