+++ etc/src/System/Etc/Internal/Resolver/Default.hs
@@ -1,5 +1,5 @@
+{-# LANGUAGE NamedFieldPuns    #-}
 {-# LANGUAGE NoImplicitPrelude #-}
-
 module System.Etc.Internal.Resolver.Default (resolveDefault) where
 
 import           RIO
@@ -13,15 +13,15 @@ import           System.Etc.Internal.Types
 
 toDefaultConfigValue :: Bool -> JSON.Value -> ConfigValue
 toDefaultConfigValue sensitive =
-  ConfigValue . Set.singleton . Default . boolToValue sensitive
+  ConfigValue . Set.singleton . Default . markAsSensitive sensitive
 
 buildDefaultResolver :: Spec.ConfigSpec cmd -> Maybe ConfigValue
 buildDefaultResolver spec =
   let resolverReducer
         :: Text -> Spec.ConfigValue cmd -> Maybe ConfigValue -> Maybe ConfigValue
       resolverReducer specKey specValue mConfig = case specValue of
-        Spec.ConfigValue def sensitive _ ->
-          let mConfigSource = toDefaultConfigValue sensitive <$> def
+        Spec.ConfigValue { Spec.defaultValue, Spec.isSensitive } ->
+          let mConfigSource = toDefaultConfigValue isSensitive <$> defaultValue
 
               updateConfig  = writeInSubConfig specKey <$> mConfigSource <*> mConfig
           in  updateConfig <|> mConfig