+++ etc/CHANGELOG.md
@@ -1,3 +1,13 @@
+0.4.0.0
+----
+**BREAKING CHANGES**
+
+* Add new `type` field to `etc/spec` with support for `string`, `number`, `bool`, `[string]`, `[number]` and `[bool]`
+* Remove `type` field in `cli` spec in favor of `type` on `etc/spec`
+* Allow ENV vars to accept supported types (only strings were allowed)
+* Allow CLI options to accept supported types (only strings and numbers were allowed)
+* Allow spec file to have array as default values
+
 0.3.2.0
 ----